requirements.txt
markdown==3.5.2
Jinja2==3.1.3
python-frontmatter==1.1.0
Pygments==2.17.2
pytest==8.0.0
